document-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
asyncpg==0.29.0
boto3==1.35.36
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
aiofiles==24.1.0
pillow==11.0.0
